Reorganize code for multi hypothesis graphs
